# docs/openapi-track6.yaml
openapi: 3.0.3
info:
  title: AMDS Track 6 API
  version: 1.0.0
  description: |
    Tenant policies, sender reputation, dynamic throughput, and campaign health.
    All endpoints require `Authorization: Bearer <AMDS_API_KEY>` unless noted.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
paths:
...
  /v1/tenants/{tenantId}/reputation/history:
    get:
      summary: Reputation score history
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 30 }
...
  /v1/campaigns/{id}/estimate:
    get:
      summary: Campaign completion ETA
      parameters:
        - name: tenant_id
          in: query
          required: true
          schema: { type: string }
        - name: recipient_count
          in: query
          required: true
          schema: { type: integer }
...